poly-cuda/31-5-24/rendering loop, some optimizations

+++ include/POLY-cuda.h
@@ -44,7 +44,7 @@ constexpr uint16_t TEXTURE_SIZE = 1024;
 constexpr uint8_t TILE_SIZE = 16;
 constexpr float AR = 1.33333f;
 constexpr uint8_t MAX_RAY_BOUNCES = 255u;
-constexpr uint8_t BVH_STACK_SIZE = 33u;
+constexpr uint16_t BVH_STACK_SIZE = 33u; // log2(MAX_TRIS * 2)
 
 // Individual tri rendering flags
 constexpr uint8_t DISABLE_RENDERING = 0x01u;
